+++ R/utils.R
@@ -119,11 +119,11 @@ pacs_base_all <- memoise::memoise(pacs_base_all_raw)
 installed_descriptions <- function(lib.loc, fields, deps = NULL) {
   installed_agg <- installed_agg_fun(lib.loc, fields)
 
-  paks <- installed_agg[, fields]
+  paks <- installed_agg[, fields, drop = FALSE]
 
   if (!is.null(deps)) {
     nams <- installed_agg[, c("Package")]
-    paks <- paks[nams %in% deps, ]
+    paks <- paks[nams %in% deps, , drop = FALSE]
   }
 
   joint_cols <- apply(paks, 1, function(x) paste(x, collapse = ","))
@@ -138,7 +138,7 @@ installed_descriptions <- function(lib.loc, fields, deps = NULL) {
     Package = unlist(lapply(seq_along(packages), function(x) replace(packages[[x]], packages[[x]] == "NA", NA))),
     stringsAsFactors = FALSE
   )
-
+browser()
   res_agg <- stats::aggregate(
     joint[, c("Version"), drop = FALSE],
     list(Package = joint$Package),
@@ -203,11 +203,11 @@ extract_deps <- function(x) {
 available_descriptions <- function(repos, fields, deps = NULL) {
   available_agg <- available_agg_fun(repos, fields)
 
-  paks <- available_agg[, fields]
+  paks <- available_agg[, fields, drop = FALSE]
 
   if (!is.null(deps)) {
     nams <- available_agg[, c("Package")]
-    paks <- paks[nams %in% deps, ]
+    paks <- paks[nams %in% deps, , drop = FALSE]
   }
 
   joint_cols <- apply(paks, 1, function(x) paste(x, collapse = ","))
@@ -218,8 +218,8 @@ available_descriptions <- function(repos, fields, deps = NULL) {
   versions <- desc_e$versions
 
   joint <- data.frame(
-    Version = unlist(sapply(seq_along(packages), function(x) replaceNA(versions[[x]], ""))),
-    Package = unlist(sapply(seq_along(packages), function(x) replace(packages[[x]], packages[[x]] == "NA", NA))),
+    Version = unlist(lapply(seq_along(packages), function(x) replaceNA(versions[[x]], ""))),
+    Package = unlist(lapply(seq_along(packages), function(x) replace(packages[[x]], packages[[x]] == "NA", NA))),
     stringsAsFactors = FALSE
   )
 
@@ -247,8 +247,7 @@ available_agg_fun_raw <- function(repos, fields) {
 available_agg_fun <- memoise::memoise(available_agg_fun_raw, cache = cachem::cache_mem(max_age = 30 * 60))
 
 expand_dependency <- function(fields) {
-  if (length(fields) == 1) {
-    stopifnot(all(fields %in% c("strong", "all", "most")))
+  if (length(fields) == 1 && fields %in% c("strong", "all", "most")) {
     switch(fields,
       strong = c("Depends", "Imports", "LinkingTo"),
       most = c("Depends", "Imports", "LinkingTo", "Suggests"),