+++ NEWS.md
@@ -1,6 +1,9 @@
 # pacs 0.4.3.9000
 
 * `tinyverse` vignette.
+* fixed edge cases for `pacs::pac_deps`.
+* two new functions `pacs::pac_deps_user` and `pacs::pac_deps_dev`.
+* updated pkgdown website.
 
 # pacs 0.4.3